% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{social}
\alias{social}
\title{Generate Social Media HTML Caption}
\usage{
social(
  github_username = "Gchism94",
  linkedIn_username = "greg-chism",
  web_link = "gregtchism.com",
  github_icon = "&#xf09b",
  linkedIn_icon = "&#xf08c",
  web_icon = "&#xf0ac",
  bg_color = "#fefefe",
  icon_color = "#006199",
  font_color = "#006199",
  font_family = "roboto"
)
}
\arguments{
\item{github_username}{GitHub username (default: "Gchism94").}

\item{linkedIn_username}{LinkedIn username (default: "greg-chism").}

\item{web_link}{Personal or professional website link (default: "gregtchism.com").}

\item{github_icon}{HTML character entity for GitHub icon (default: "&#xf09b").}

\item{linkedIn_icon}{HTML character entity for LinkedIn icon (default: "&#xf08c").}

\item{web_icon}{HTML character entity for a generic web icon (default: "&#xf0ac").}

\item{bg_color}{Background color for the HTML div (default: "#fefefe").}

\item{icon_color}{Color for the icons (default: "#006199").}

\item{font_color}{Color for the text (default: "#006199").}

\item{font_family}{Font family for the text (default: "roboto").}
}
\value{
An HTML string representing the formatted social media details.
}
\description{
This function creates an HTML string to display social media icons and
user details. It is designed to be used in environments that can render
HTML content such as R Markdown documents or Shiny applications.
}
\examples{
social(github_username = "YourGitHub",
       linkedIn_username = "YourLinkedIn",
       web_link = "YourWebsite.com")
}
